Sync branch structure with latest updates
